# backend/docs/swagger.yaml
basePath: /
definitions:
  controllers.AuthSigninReq:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  controllers.ErrorResponse:
    properties:
      message:
        type: string
    required:
    - message
    type: object
  controllers.PostNewReq:
    properties:
      body:
        type: string
      title:
        maxLength: 100
        type: string
    required:
    - body
    - title
    type: object
  controllers.UserCreateReq:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  entities.Post:
    properties:
      body:
        type: string
      created_at:
        format: date-time
        type: string
      id:
        type: integer
      title:
        type: string
      updated_at:
        format: date-time
        type: string
      user_id:
        type: integer
      username:
        type: string
    required:
    - body
    - created_at
    - id
    - title
    - updated_at
    - user_id
    - username
    type: object
  entities.User:
    properties:
      id:
        type: integer
      username:
        type: string
    required:
    - id
    - username
    type: object
host: localhost:9000
info:
  contact: {}
  description: このswaggerはweb application演習5班のAPI仕様書です。
  title: web-application team 5 API
  version: "1.0"
paths:
  /posts:
    get:
      consumes:
      - application/json
      description: 投稿一覧を取得します
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.Post'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: get list posts
      tags:
      - posts
    post:
      consumes:
      - application/json
      description: サインインしているユーザーで、指定されたタイトル、本文の投稿を作成する
      parameters:
      - description: リクエストパラメータ
        in: body
        name: signinPost
        required: true
        schema:
          $ref: '#/definitions/controllers.PostNewReq'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: new post API
      tags:
      - posts
  /posts/{postID}:
    get:
      consumes:
      - application/json
      description: 投稿の詳細を取得します
      parameters:
      - description: Post ID
        in: path
        name: postID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Post'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: get post detail
      tags:
      - posts
  /signin:
    post:
      consumes:
      - application/json
      description: usernameとpasswordでsigninします。
      parameters:
      - description: リクエストパラメータ
        in: body
        name: signinPost
        required: true
        schema:
          $ref: '#/definitions/controllers.AuthSigninReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: singin API
      tags:
      - auth
  /users:
    post:
      consumes:
      - application/json
      description: usernameとpasswordでuserを作成します。
      parameters:
      - description: リクエストパラメータ
        in: body
        name: userCreate
        required: true
        schema:
          $ref: '#/definitions/controllers.UserCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Create User API
      tags:
      - users
  /users/me:
    get:
      description: ログインしているユーザーの情報を取得します
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: get user me
      tags:
      - users
swagger: "2.0"